Split movement into player control and generic movement
